ndg_manpage

Module escape

Statics

ROFF_ESCAPES

Map of characters that need to be escaped in manpages

TROFF_ESCAPE

TROFF_FORMATTING

Functions

escape_leading_dot

Escape a leading dot to prevent it from being interpreted as a troff command

escape_non_macro_lines

Escape lines except those starting with man macros we emit (e.g., .IP)

man_escape

Escapes a string for use in manpages